% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDetect_BJ.R
\name{BJ_threshold_calc}
\alias{BJ_threshold_calc}
\title{Simulating the threshold for the IDetect Berk-Jones procedure}
\usage{
BJ_threshold_calc(n, points, N, alpha, mbj = FALSE)
}
\arguments{
\item{n}{The length of the p-value sequence}

\item{points}{The length of the smallest interval considered (step parameter)}

\item{N}{The number of repetitions}

\item{alpha}{The upper quantile to be used}

\item{mbj}{Boolean, if `TRUE`, the modified Berk-Jones statistic is used,
otherwise the standard form}
}
\value{
Returns the estimated threshold value
}
\description{
Calculate the threshold to be used by `idetect.th_BJ_re` or `idetect.th_BJ_le`.
The threshold is calculated as an upper quantile of the sequence of test statistics,
 such that the procedure under the null (for uniform p-values) does not detect any change-points.
}
