% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CP_loc_alpha05}
\alias{CP_loc_alpha05}
\title{Change-point locations for alpha = 1/2, for the Gaussian model}
\format{
An object of class \code{data.frame} with 31 rows and 12 columns.
}
\usage{
CP_loc_alpha05
}
\description{
This data contains the locations of the true change-points
for the DOS statistic with alpha = 1, under the Gaussian model,
for various values of mu (columns) and pi1 (rows).
This data is obtained using Mathematica.
}
\keyword{datasets}
