% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dos_functions.R
\name{dos_fun}
\alias{dos_fun}
\title{DOS sequence calculation}
\usage{
dos_fun(p_seq, alpha = 1, exc = 0)
}
\arguments{
\item{p_seq}{A sequence of p-values}

\item{alpha}{DOS sequence power parameter}

\item{exc}{Number of excluded values from the beginning}
}
\value{
A list containing
\item{dos_seq}{the DOS sequence}
\item{cp_loc}{the location of the maximum of the DOS sequence}
}
\description{
This function calculates the DOS sequence and the estimated change-point location
}
\examples{
p_seq <- sim_pval(100, 0, 3, 0.1, 0)
dos_fun(p_seq)
}
