% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estimable_prop_slope}
\alias{estimable_prop_slope}
\title{Estimable proportion for alpha = 1, for the Gaussian model}
\format{
An object of class \code{data.frame} with 31 rows and 17 columns.
}
\usage{
estimable_prop_slope
}
\description{
This data contains the values corresponding to the part of the
false null proportion that the DOS method with alpha = 1
is able to estimate asymptotically.
This data is obtained using Mathematica.
}
\keyword{datasets}
