% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions_functions.R
\name{hd_est}
\alias{hd_est}
\title{Hoang & Dickhaus randomised p-values calculation}
\usage{
hd_est(p_seq, lambda = 0.5, c = NULL)
}
\arguments{
\item{p_seq}{A sequence of p-values}

\item{lambda}{Parameter for Storey's estimator and for randomisation}

\item{c}{Threshold parameter for calculating the randomised p-values.
If NULL, the threshold is calculated by optimising a function defined in Eq. 8 in the paper.}
}
\value{
A list containing
* `random_p_seq` the sequence of randomised p-values

* `pi1_est` the corresponding Storey's estimator of the false null proportion based on the randomised sequence and for given lamdba
}
\description{
This function produces a sequence of randomised p-values
using the method proposed in
Hoang, A.-T., & Dickhaus, T. (2020).
On the usage of randomized p-values in the Schweder–Spjøtvoll estimator.
Annals of the Institute of Statistical Mathematics, 74, 289–319.
}
