% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{Hwang_est}
\alias{Hwang_est}
\title{Hwang et al. Slope Difference estimator calculation}
\usage{
Hwang_est(p_seq)
}
\arguments{
\item{p_seq}{A sequence of p-values}
}
\value{
Estimated true null proportion
}
\description{
This function calculates the estimates of the true null proportion using the
Slope Difference estimator proposed in Hwang, Y. T. et al. (2014).
}
\examples{
p_seq <- sim_pval(100, 0, 3, 0.1, 0)
Hwang_est(p_seq)
}
\references{
Hwang, Y. T., Kuo, H. C., Wang,
 C. C., & Lee, M. F. (2014). Estimating the number of true null hypotheses
  in multiple hypothesis testing. Statistics and Computing, 24(3), 399–416.
  https://doi.org/10.1007/s11222-013-9377-5
}
