% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDetect_BJ.R
\name{ID_BJ_group}
\alias{ID_BJ_group}
\title{Local FDR estimator using IDetect Berk-Jones procedure}
\usage{
ID_BJ_group(
  p_seq,
  thr = NULL,
  points = 5,
  mode = "left",
  mbj = FALSE,
  plot = FALSE,
  alpha = 0.05,
  N = 1000
)
}
\arguments{
\item{p_seq}{The sequence of sorted p-values}

\item{thr}{The threshold for the procedure. If `NULL`, then calculate using the
`BJ_threshold_calc` function.}

\item{points}{The length of the smallest interval considered (step parameter)}

\item{mode}{Set to 'left exp' for the procedure with left expanding intervals, or 'right exp'
for the procedure with right expanding intervals}

\item{mbj}{Boolean, if `TRUE`, the modified Berk-Jones statistic is used,
otherwise the standard form}

\item{plot}{Boolean, if `TRUE` a plot of the estimated local FDR will be produced}

\item{alpha}{The significance level for calculating the threshold}

\item{N}{The number of repetitions for calculating the threshold}
}
\value{
Returns a list of 2 elements
\item{lfdr}{The estimated lfdr values corresponding to the p-values sequence}
\item{groups}{A dataframe with the information about the resulting grouping of the p-values
based on their significance}
}
\description{
Calculate the estimate of the local FDR using the IDetect Berk-Jones procedure
with right or left expanding intervals
}
