% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDetect_BJ.R
\name{idetect.th_BJ}
\alias{idetect.th_BJ}
\title{IDetect method with the Berk-Jones statistic}
\usage{
idetect.th_BJ(
  p_seq,
  thr = NULL,
  points = 5,
  mode = "left",
  mbj = FALSE,
  alpha = 0.05,
  N = 1000
)
}
\arguments{
\item{p_seq}{The sorted sequence of p-values}

\item{thr}{The threshold for the Berk-Jones statistic. If `NULL` the threshold is calculated through simulations}

\item{points}{The number of points added to the interval at each step (the minimum number of points)}

\item{mode}{'left' or 'right' to indicate the type of expanding intervals to consider:
'left' for left-expanding intervals and 'right' for right-expanding intervals}

\item{mbj}{Boolean, if `TRUE`, the modified Berk-Jones statistic is used,
otherwise the standard form}

\item{alpha}{The significance level for calculating the threshold}

\item{N}{The number of repetitions for calculating the threshold}
}
\value{
Returns a list of 2 elements
\item{changepoint}{The locations of the estimated change-points}
\item{full_information}{A dataframe with the information about the Berk-Jones statistic on the subintervals
at each step of the algorithm.
The endpoints of the interval considered are in the first two columns, the location of the maximum in the third,
and the value of the Berk-Jones statistic at the fourth column}
}
\description{
Given a sequence of sorted p-values, perform segmentation into uniform segments using
the IDetect method with Berk-Jones statistic.
}
\references{
Li, J., & Siegmund, D. (2015). Higher criticism:
 $p$-values and criticism. The Annals of Statistics, 43(3), 1323–1350.
  https://doi.org/10.1214/15-aos1312
}
\seealso{
[BJ_threshold_calc()]
}
