% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDetect_BJ.R
\name{idetect.th_BJ_re}
\alias{idetect.th_BJ_re}
\title{IDetect method with right-expanding intervals and the Berk-Jones statistic}
\usage{
idetect.th_BJ_re(
  p_seq,
  thr,
  points = 5,
  s = 1,
  e = length(p_seq),
  k_r = 1,
  mbj = FALSE
)
}
\arguments{
\item{p_seq}{The sorted sequence of p-values}

\item{thr}{The threshold for the Berk-Jones statistic}

\item{points}{The number of points added to the interval at each step (the minimum number of points)}

\item{s}{The starting point of the current interval observed for change (for the recursion)}

\item{e}{The endpoint of the current interval observed for change (for the recursion)}

\item{k_r}{The current step in expanding the right edge of the subintervals within the (s,e) range.}

\item{mbj}{Boolean, if `TRUE`, the modified Berk-Jones statistic is used,
otherwise the standard form}
}
\value{
Returns a list of 2 elements
\item{changepoint}{The locations of the estimated change-points}
\item{full_information}{A dataframe with the information about the Berk-Jones statistic on the subintervals
at each step of the algorithm.
The endpoints of the interval considered are in the first two columns, the location of the maximum in the third,
and the value of the Berk-Jones statistic at the fourth column}
}
\description{
Given a sequence of sorted p-values, perform segmentation into uniform segments using
the IDetect method with Berk-Jones statistic. We consider right-expanding intervals.
}
