% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{LLF_Grenander_est}
\alias{LLF_Grenander_est}
\title{Langaas, Lindqvist & Ferkingstad estimator calculation}
\usage{
LLF_Grenander_est(p_seq)
}
\arguments{
\item{p_seq}{A sequence of p-values}
}
\value{
Estimated true null proportion
}
\description{
Estimator of the true null proportion at the longest constant interval in the Grenander estimator,
as proposed in Langaas, M., Lindqvist, B. H., & Ferkingstad, E. (2005).
Estimating the Proportion of True Null Hypotheses, with Application to DNA Microarray Data.
 Journal of the Royal Statistical Society. Series B (Statistical Methodology), 67(4), 555–572.
}
\examples{
p_seq <- sim_pval(100, 0, 3, 0.1, 0)
LLF_Grenander_est(p_seq)

}
\references{
Langaas, M., Lindqvist, B. H., & Ferkingstad, E. (2005).
Estimating the Proportion of True Null Hypotheses, with Application to DNA Microarray Data.
 Journal of the Royal Statistical Society. Series B (Statistical Methodology), 67(4), 555–572.
 https://doi.org/10.1111/j.1467-9868.2005.00515.x
}
