% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDetect_BJ.R
\name{max_BJ}
\alias{max_BJ}
\title{Berk-Jones statistic on subsets}
\usage{
max_BJ(ind = c(1, length(p_seq)), p_seq, mbj = FALSE)
}
\arguments{
\item{ind}{A vector containing the starting and the ending point of the interval}

\item{p_seq}{The sorted sequence of p-values}

\item{mbj}{Boolean, if `TRUE`, the modified Berk-Jones statistic is used,
otherwise the standard form}
}
\value{
Returns a list of 2 elements
\item{val}{the value of the Berk-Jones statistic}
\item{loc}{the location in the corresponding sequence where the maximum is achieved}
}
\description{
Given a sequence of sorted p-values and the start and endpoint indices \eqn{s} and \eqn{e},
check if the values \eqn{p_{(s)},..., p_{(e)}} form a sample from uniform \eqn{U[p_{(s-1)}, p_{(e+1)}]}
distribution.
}
\references{
Li, J., & Siegmund, D. (2015). Higher criticism:
 $p$-values and criticism. The Annals of Statistics, 43(3), 1323–1350.
  https://doi.org/10.1214/15-aos1312
}
