% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{MGF_est}
\alias{MGF_est}
\title{This function calculates the estimates of the true null proportion using the method 'MGF'
proposed in Broberg, P. (2005). A comparative review of estimates of the proportion
unchanged genes and the false discovery rate. BMC Bioinformatics, 6, 199. The code belongs to Per Broberg and
it was a part of the now deprecated package 'SAGx' in BioConductor. 'SAGx' can be downloaded from
https://www.bioconductor.org/packages//2.10/bioc/html/SAGx.html}
\usage{
MGF_est(p_seq)
}
\arguments{
\item{p_seq}{A sequence of p-values}
}
\value{
Estimated true null proportion
}
\description{
This function calculates the estimates of the true null proportion using the method 'MGF'
proposed in Broberg, P. (2005). A comparative review of estimates of the proportion
unchanged genes and the false discovery rate. BMC Bioinformatics, 6, 199. The code belongs to Per Broberg and
it was a part of the now deprecated package 'SAGx' in BioConductor. 'SAGx' can be downloaded from
https://www.bioconductor.org/packages//2.10/bioc/html/SAGx.html
}
