% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{MR_est}
\alias{MR_est}
\title{Meinshausen & Rice estimator calculation}
\usage{
MR_est(p_seq, alpha = 0.05)
}
\arguments{
\item{p_seq}{A sequence of p-values}

\item{alpha}{For the confidence interval of level 1-alpha}
}
\value{
Estimated false null proportion
}
\description{
This function calculates the false null proportion estimates
using the method proposed in Meinshausen, N., & Rice, J. (2006)
}
\examples{
p_seq <- sim_pval(100, 0, 3, 0.1, 0)
MR_est(p_seq)
}
\references{
"Meinshausen, N., & Rice, J. (2006). Estimating the proportion of false null
 hypotheses among a large number of independently tested hypotheses.
 Annals of Statistics, 34(1), 373–393" https://doi.org/10.1214/009053605000000741
}
