% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dos_functions.R
\name{sim_pval}
\alias{sim_pval}
\title{Generating p-values from the Gaussian model}
\usage{
sim_pval(n, null_mean, alt_mean, pi1, rho)
}
\arguments{
\item{n}{The length of the output vector of p-values}

\item{null_mean}{A vector of mean values under the null, or a single
value if it is the same for all variables}

\item{alt_mean}{A vector of mean values under the alternative, or a single
value if it is the same for all variables}

\item{pi1}{A proportion of false null p-values}

\item{rho}{A correlation parameter between 0 and 1, such that the correlation between
each pair of p-values is equal to rho}
}
\value{
A vector of p-values
}
\description{
This function generates one-sided p-values from the Gaussian mean testing
where the p-values are calculated assuming that the statistics
have standard normal distribution under the null
and a mean greater than zero under the alternative.
}
