% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dos_functions.R
\name{storey_pi1est}
\alias{storey_pi1est}
\title{Storey's estimator function}
\usage{
storey_pi1est(lambda, p_seq, mod = FALSE)
}
\arguments{
\item{lambda}{A vector of tuning parameter values}

\item{p_seq}{A sequence of p-values}

\item{mod}{Whether a modified Storey's estimator is to be calculated (see Blanchard, G.,
& Roquain, É. (2009). Adaptive false discovery rate control under independence and dependence.
 Journal of Machine Learning Research, 10, 2837–2871.)}
}
\value{
A list containing
* `lambda` Input vector lambda

* `est` The corresponding vector of estimates
}
\description{
This function calculates the Storey's estimator
of the false null proportion, given the tuning parameter values
and the sequence of p-values
}
