% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{turkheimer_est}
\alias{turkheimer_est}
\title{Turkheimer et al. Change-point based estimator}
\usage{
turkheimer_est(p_seq, alpha)
}
\arguments{
\item{p_seq}{A sequence of p-values}

\item{alpha}{Significance level for iterative uniformity testing}
}
\value{
Estimated true null proportion
}
\description{
This function calculates the estimates of the true null proportion using the
Change-point based estimator proposed in
 Turkheimer, F. E. et al. (2001)
}
\examples{
p_seq <- sim_pval(100, 0, 3, 0.1, 0)
turkheimer_est(p_seq, 0.05)
}
\references{
Turkheimer, F. E., Smith, C. B., & Schmidt, K. (2001).
 Estimation of the number of “true” null hypotheses in multivariate
  analysis of neuroimaging data. NeuroImage, 13(5), 920–930.
  https://doi.org/10.1006/nimg.2001.0764
}
