% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions_functions.R
\name{bh_adaptive}
\alias{bh_adaptive}
\title{Adaptive Benjamini-Hochberg procedure}
\usage{
bh_adaptive(p_seq, pi0, alpha, sig)
}
\arguments{
\item{p_seq}{A sequence of p-values}

\item{pi0}{Estimated true null proportion}

\item{alpha}{FDR control level}

\item{sig}{A vector matching p-values, zero if the corresponding p-value
is true null, and 1 if it is false null}
}
\value{
A list containing
* `fdr` the realised proportion of false discoveries among all discoveries

* `power` the proportion of true discoveries among all false null p-values
}
\description{
This function calculates the FDR and the power of the adaptive BH procedure
for a given value of the proportion estimator
(Benjamini, Y., & Hochberg, Y. (2000). On the Adaptive Control of the False
Discovery Rate in Multiple Testing with Independent Statistics.
 Journal of Educational and Behavioral Statistics, 25(1), 60.)
}
