% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dos_functions.R
\name{dos_fun}
\alias{dos_fun}
\title{DOS sequence calculation}
\usage{
dos_fun(p_seq, alpha = 1, exc = 0)
}
\arguments{
\item{p_seq}{A sequence of p-values}

\item{alpha}{DOS sequence power parameter}

\item{exc}{Number of excluded values from the beginning}
}
\value{
A list containing
* `dos_seq` the DOS sequence

* `cp_loc` the location of the maximum of the DOS sequence
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
