% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{JD_est}
\alias{JD_est}
\title{Jiang & Doerge estimator calculation}
\usage{
JD_est(p_seq, N)
}
\arguments{
\item{p_seq}{A sequence of p-values}

\item{N}{The number of bootstrap samples}
}
\value{
Estimated true null proportion
}
\description{
This function calculates the true null proportion estimates
using the method proposed in Jiang, H., & Doerge, R. W. (2008)
}
\examples{
p_seq <- sim_pval(100, 0, 3, 0.1, 0)
JD_est(p_seq, 1000)
}
\references{
Jiang, H., & Doerge, R. W. (2008).
Estimating the Proportion of True Null Hypotheses
for Multiple Comparisons. Cancer Informatics, 6, 25--32
https://doi.org/10.1177/117693510800600001
}
