% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{LLF_Grenander_est}
\alias{LLF_Grenander_est}
\title{Estimator of the true null proportion at the longest constant interval in the Grenander estimator,
as proposed in Langaas, M., Lindqvist, B. H., & Ferkingstad, E. (2005).
Estimating the Proportion of True Null Hypotheses, with Application to DNA Microarray Data.
 Journal of the Royal Statistical Society. Series B (Statistical Methodology), 67(4), 555–572.}
\usage{
LLF_Grenander_est(p_seq)
}
\arguments{
\item{p_seq}{A sequence of p-values}
}
\value{
Estimated true null proportion
}
\description{
Estimator of the true null proportion at the longest constant interval in the Grenander estimator,
as proposed in Langaas, M., Lindqvist, B. H., & Ferkingstad, E. (2005).
Estimating the Proportion of True Null Hypotheses, with Application to DNA Microarray Data.
 Journal of the Royal Statistical Society. Series B (Statistical Methodology), 67(4), 555–572.
}
