% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{PRE_est}
\alias{PRE_est}
\title{Broberg's PRE estimator calculation}
\usage{
PRE_est(p_seq)
}
\arguments{
\item{p_seq}{A sequence of p-values}
}
\value{
Estimated true null proportion
}
\description{
This function calculates the estimates of the true null proportion using the method 'MGF'
proposed in Broberg, P. (2005). A comparative review of estimates of the proportion
unchanged genes and the false discovery rate. BMC Bioinformatics, 6, 199. The code belongs to Per Broberg and
it was a part of the now deprecated package 'SAGx' in BioConductor. 'SAGx' can be downloaded from
https://www.bioconductor.org/packages//2.10/bioc/html/SAGx.html
}
\examples{
p_seq <- sim_pval(100, 0, 3, 0.1, 0)
PRE_est(p_seq)
}
\references{
Broberg, P. (2005). A comparative review of estimates of the proportion
unchanged genes and the false discovery rate. BMC Bioinformatics, 6, 199.
https://doi.org/10.1186/1471-2105-6-199
}
